增加redis 的查询功能

+++ src/main/java/com/item_backend/model/pojo/PageResult.java
@@ -15,18 +15,18 @@ import java.util.List;
 public class PageResult<T> {
 
 
-    private Long total; // 数据总数
+    private Integer total; // 数据总数
     private List<T> rows; // 数据
     private String message;
     private Integer code; // 返回状态码
 
-    public PageResult(Long total, List<T> rows) {
+    public PageResult(Integer total, List<T> rows) {
         super();
         this.total = total;
         this.rows = rows;
     }
 
-    public PageResult(Long total, List<T> rows, String message, Integer code) {
+    public PageResult(Integer total, List<T> rows, String message, Integer code) {
         this.total = total;
         this.rows = rows;
         this.message = message;
